cleaned up alot of code, took out all my functions for future value
