+++ client/src/pages/RetirementPage/RetirementPage.tsx
@@ -14,23 +14,10 @@ import { editRetireGoalTitle } from "../../redux/features/retirementSlice";
 import RetirementSummary from "./RetirementSummary";
 import RetirementExplain from "./RetirementExplain";
 import MenuIcon from "@mui/icons-material/Menu";
-
-interface SelectedGoalType {
-  type: string;
-  id: string;
-  age: {
-    currentAge: number;
-    retireAge: number;
-    lifeExpectancy: number;
-  };
-  savings: number;
-  monthlyContribution: number;
-  budget: number;
-  preRate: number;
-  postRate: number;
-  inflation: number;
-  title: string;
-}
+import futureValueWhatYouHave from "../../components/helperFunctions/futureValueWhatYouHave";
+import getWhatYouNeedFinalPrice from "../../components/helperFunctions/getWhatYouNeedFinalPrice";
+import futureValueWhatYouWillNeed from "../../components/helperFunctions/futureValueWhatYouWillNeed";
+import { getMonthlyPaymentForHave, getMonthlyPaymentForNeed } from "../../components/helperFunctions/getMonthlyPaymentForHave";
 
 interface AgeNum {
   age: number;
@@ -116,6 +103,13 @@ export default function RetirementPage() {
   // Show State
   const [show, setShow] = React.useState(true);
 
+  //
+  // const [open, setOpen] = React.useState(false);
+
+  // const toggleDrawer = (newOpen: boolean) => () => {
+  //   setOpen(newOpen);
+  // };
+
   // Final Price State
   const [needFinalPrice, setNeedFinalPrice] = React.useState<number>(0);
 
@@ -132,256 +126,15 @@ export default function RetirementPage() {
     setSaveBtn(false);
   };
 
-  // Future Value Function ... returns array
+  //* UseEffect here handles all the math needed for charts
   // link explains that investments and loans typically compound monthly ... so all we do is take the nomial rate / coumponding periods === 12
   React.useEffect(() => {
-    // Put my foor loops in a useEffect ... my data was getting ran too much outside of it
-    function futureValueWhatYouHave(obj: SelectedGoalType) {
-      const {
-        age: { currentAge, retireAge },
-        savings,
-        monthlyContribution,
-        preRate,
-      } = obj;
-
-      const newPreRate = preRate / 100;
-      let res = [];
-      for (let i = 0; i <= retireAge - currentAge; i++) {
-        //Without making sure i > 0 ... the for loop stopped 1 year short ... now its stops at the right retirement age
-        if (i > 0) {
-          let age = currentAge;
-          const amount = savings;
-          const time = i;
-          const monthlyP = monthlyContribution;
-
-          const months = 12;
-          const rate = newPreRate / months;
-          //const c = obj.compound;
-          //const equivalentPeriodicRate = Math.pow(1 + rate, c / months) - 1;
-          const fvOfPv = amount * Math.pow(1 + rate, time * months);
-
-          const top = Math.pow(1 + rate, time * 12) - 1;
-          //console.log(top)
-          const value = monthlyP * (top / rate);
-
-          res.push({
-            age: i === 1 ? currentAge : age + i,
-            value: value + fvOfPv,
-          });
-        }
-      }
-      const highestNum = Math.max(...res.map((item) => item.value));
-
-      return {
-        data: res,
-        highestNum: USDollar.format(highestNum),
-        highestNumNoFormat: highestNum,
-      };
-    }
-
-    function getWhatYouNeedFinalPrice(obj: SelectedGoalType) {
-      const {
-        age: { retireAge, lifeExpectancy },
-        budget,
-        postRate,
-        inflation,
-      } = obj;
-      const newPostRate = postRate / 100;
-      const newInflation = inflation / 100;
-      let res = [];
-      for (let i = 0; i < lifeExpectancy - retireAge; i++) {
-        const age = retireAge;
-
-        const time = i + 1;
-
-        const monthlyP = budget;
-
-        const months = 12;
-        const addInflationAndPostRate = (1 + newPostRate) / (1 + newInflation) - 1;
-
-        const rate = addInflationAndPostRate / months;
-        //const equivalentPeriodicRate = Math.pow(1 + rate, c / months) - 1;
-
-        const toThePowerTop = Math.pow(1 + rate, -time * months);
-        const subtractToThePower = 1 - toThePowerTop;
-        const top = monthlyP * subtractToThePower;
-        const value = top / rate;
-
-        // if inflation and postRate === each ohter it means we have a nomial rate of 0
-        const yearlyDeposit = monthlyP * 12;
-        //console.log(yearlyDeposit)
-        const rateOfZero = yearlyDeposit * time;
-
-        res.push({
-          age: age + i,
-          value: newInflation === newPostRate ? rateOfZero : value,
-        });
-      }
-
-      const highestNum = Math.max(...res.map((item) => item.value));
-      return highestNum;
-    }
-
-    function futureValueWhatYouWillNeed(obj: SelectedGoalType) {
-      const {
-        age: { currentAge, retireAge },
-
-        postRate,
-        inflation,
-      } = obj;
-      const newPostRate = postRate / 100;
-      const newInflation = inflation / 100;
-      // This function grabs the monthy payment ... becasue I need it in the foor loop below to create my what you need chart
-      //* We simply solve for PMT
-      //Future Value of an Ordinary Annuity
-      //* FV = PMT * ((1 + rate/months)^ time * months - 1) / (rate / 12)
-      function getMonthlyPayment() {
-        let monthlyContribution = 0;
-        for (let i = 0; i < retireAge - currentAge; i++) {
-          const time = retireAge - currentAge;
-
-          const moneyNeededForBudgetNumber = getWhatYouNeedFinalPrice(obj);
-          const addInflationAndPostRate = (1 + newPostRate) / (1 + newInflation) - 1;
-          //console.log(addInflationAndPostRate)
-          const months = 12;
-          const rate = addInflationAndPostRate / months;
-
-          const toThePowerTop = Math.pow(1 + rate, time * months) - 1;
-          const topAndBottom = toThePowerTop / rate;
-
-          const PMT = moneyNeededForBudgetNumber / topAndBottom;
-
-          // if inflation and postRate === each ohter it means we have a nomial rate of 0
-          const numberOfPayments = time * months;
-          const pmtZeroInterest = moneyNeededForBudgetNumber / numberOfPayments;
-
-          monthlyContribution = newInflation === newPostRate ? pmtZeroInterest : PMT;
-        }
-        return monthlyContribution;
-      }
-
-      // Here we take the retire age - age to get the length
-      //* and just loop ... each loop/year is the future value of the concsective year and it keeps getting bigger as the years increase
-      // Future Value of a Savings Annuity = PV(1 + (rate/months))^ time * months + (PMT * PMT * ((1 + rate/months)^ time * months - 1) / (rate / 12))
-      let res = [];
-      for (let i = 0; i <= retireAge - currentAge; i++) {
-        if (i > 0) {
-          let age = currentAge;
-
-          const amount = 0;
-          const time = i;
-          const monthlyP = getMonthlyPayment();
-          //const c = obj.compound;
-          const months = 12;
-          const addInflationAndPostRate = (1 + newPostRate) / (1 + newInflation) - 1;
-
-          const rate = addInflationAndPostRate / months;
-          //const equivalentPeriodicRate = Math.pow(1 + rate, c / months) - 1;
-          const fvOfPv = amount * Math.pow(1 + rate, time * months);
-          const top = Math.pow(1 + rate, time * 12) - 1;
-          const value = monthlyP * (top / rate);
-
-          // if inflation and postRate === each ohter it means we have a nomial rate of 0
-          const yearlyDeposit = monthlyP * 12;
-          const rateOfZero = yearlyDeposit * time;
-
-          res.push({
-            age: i === 1 ? currentAge : age + i,
-            value: newInflation === newPostRate ? rateOfZero : value + fvOfPv,
-          });
-        }
-      }
-      const highestNum = Math.max(...res.map((item) => item.value));
-
-      return {
-        data: res,
-        highestNum: USDollar.format(highestNum),
-        highestNumNoFormat: highestNum,
-      };
-    }
-
-    // Payout Annutiy Function is being used with inflation to calculate monthly withdraw
-    function getMonthlyPaymentHave(obj: SelectedGoalType, highNum: number) {
-      const {
-        age: { retireAge, lifeExpectancy },
-
-        postRate,
-        inflation,
-      } = obj;
-      const newPostRate = postRate / 100;
-      const newInflation = inflation / 100;
-
-      const time = lifeExpectancy - retireAge;
-
-      const finalPrice = highNum;
-      const addInflationAndPostRate = (1 + newPostRate) / (1 + newInflation) - 1;
-      //console.log(addInflationAndPostRate)
-      const months = 12;
-      const rate = addInflationAndPostRate / months;
-
-      const toThePowerTop = Math.pow(1 + rate, -time * months);
-      const subtractToThePower = 1 - toThePowerTop;
-      const topDivideBottom = subtractToThePower / rate;
-      const value = finalPrice / topDivideBottom;
-
-      // This for PMT with not adjusted with infaltion
-      const rateNoInflation = newPostRate / months;
-
-      const toThePowerTopNoInflation = Math.pow(1 + rateNoInflation, -time * months);
-      const subtractToThePowerNoInflation = 1 - toThePowerTopNoInflation;
-      const topDivideBottomNoInflation = subtractToThePowerNoInflation / rateNoInflation;
-      const valueNoInflation = finalPrice / topDivideBottomNoInflation;
-
-      // if inflation and postRate === each ohter it means we have a nomial rate of 0
-      const numberOfPayments = time * months;
-      const pmtZeroInterest = finalPrice / numberOfPayments;
-
-      return {
-        value: newInflation === newPostRate ? pmtZeroInterest : value,
-        valueNoInflation,
-      };
-    }
-
-    // Payout Annutiy for getting montly contribution for what you need final price
-    function getMonthlyPaymentNeed(obj: SelectedGoalType) {
-      let monthlyContribution = 0;
-      const {
-        age: { retireAge, currentAge },
-
-        postRate,
-        inflation,
-      } = obj;
-      const newPostRate = postRate / 100;
-      const newInflation = inflation / 100;
-
-      const time = retireAge - currentAge;
-
-      const moneyNeededForBudgetNumber = getWhatYouNeedFinalPrice(obj);
-      const addInflationAndPostRate = (1 + newPostRate) / (1 + newInflation) - 1;
-
-      const months = 12;
-      const rate = addInflationAndPostRate / months;
-
-      const toThePowerTop = Math.pow(1 + rate, time * months) - 1;
-      const topAndBottom = toThePowerTop / rate;
-
-      const PMT = moneyNeededForBudgetNumber / topAndBottom;
-
-      // if inflation and postRate === each ohter it means we have a nomial rate of 0
-      const numberOfPayments = time * months;
-      const pmtZeroInterest = moneyNeededForBudgetNumber / numberOfPayments;
-
-      monthlyContribution = newInflation === newPostRate ? pmtZeroInterest : PMT;
-
-      return monthlyContribution;
-    }
-
     if (selectedGoal) {
       setHave(futureValueWhatYouHave(selectedGoal));
       setNeedFinalPrice(getWhatYouNeedFinalPrice(selectedGoal));
       setNeed(futureValueWhatYouWillNeed(selectedGoal));
-      setHaveRetireBudget(getMonthlyPaymentHave(selectedGoal, futureValueWhatYouHave(selectedGoal).highestNumNoFormat));
-      setNeedMonthlyContribution(getMonthlyPaymentNeed(selectedGoal));
+      setHaveRetireBudget(getMonthlyPaymentForHave(selectedGoal, futureValueWhatYouHave(selectedGoal).highestNumNoFormat));
+      setNeedMonthlyContribution(getMonthlyPaymentForNeed(selectedGoal));
     }
   }, [selectedGoal, retireGoals, dispatch]); // eslint-disable-line
